Fix enum construction from integers.
Construction of enum from integers didn't work as exepcted when the
arguments weren't constants.

Assuming E in an enum type, with `E(x)` the argument `x` can now be of
any integer type, and both constant and non-constant. However, `x` may
not be larger than the maximum int64 value (because that's our internal
enum representation). If it is, that will trigger a runtime error.